backend/requirements.txt
easyocr==1.7.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0 